Calculate rating with db query rather than in code
Should in the end be calculated periodically and cached though

+++ src/main/kotlin/io/aesy/yumme/dto/RatingDto.kt
@@ -8,5 +8,5 @@ data class RatingDto(
     var average: Double?,
 
     @JsonProperty
-    var count: Int?
+    var count: Long?
 )
